The registration path is working
There are no tests and there is no way to log out but the registration
ceremony is working.  As far as tests, I looked at the previous demo
code and they seem more interested in testing the valid path which
doesn't interest me at all.  I'm interested in all of the failure
paths and making sure the user is properly informed.